Updates to support adding ACIS DEA HK and Ephemeris pseudo-MSIDs
- Add L0 (definitive) ephemeris data.
- Update converters to use more consistent prefix convention and include L0 (definitive) ephemeris.
- Update docs: Pseudo-MSIDs doc and standalone fetch tutorial.
- Much improved NOTES.add content.
- Improve fix_ingest.py:
   - Supply options by optparse
   - Allow for dry-run
   - Allow for partial overwrite to handle Ephemeris files
- Ingest_fits_archive.py can handle ephemeris files where no time-based query is available.
- Changes to update_archive.py:
  - If a file overlaps with a previous archive file then overwrite the old values with
    the new.  This is done by setting appropriate bad (quality) values in the TIME column.
  - Filetype-sensitive max-gap default value (601 for ephem, 66 for DEA HK, 32.9 otherwise)
  - New logic in get_archive_files to deal with ephemeris files for which time-based
    archive queries are not supported.

+++ fix_ingest_h5.py
@@ -7,7 +7,9 @@ retrieved by the archive.
 import sys
 import os
 import shutil
+import optparse
 
+import numpy
 import tables
 from Chandra.Time import DateTime
 import Ska.DBI
@@ -15,9 +17,21 @@ import Ska.Table
 
 from context_def import datadir, ft, files
 
+def get_options():
+    parser = optparse.OptionParser()
+    parser.add_option("--dry-run",
+                      action="store_true",
+                      help="Dry run (no actual file or database updatees)")
+    parser.add_option("--overwrite-partial",
+                      action="store_true",
+                      help="Overwrite the overlapping bits of file (e.g. for EPHEM)")
+    return parser.parse_args()
+
+opt, args = get_options()
+
 filetypes = Ska.Table.read_ascii_table('filetypes.dat')
-if len(sys.argv) == 2:
-    filetypes = filetypes[ filetypes['content'] == sys.argv[1].upper() ]
+if len(args) > 0:
+    filetypes = filetypes[ filetypes['content'] == args[0].upper() ]
 
 for filetype in filetypes:
     # Update attributes of global ContextValue "ft".  This is needed for
@@ -34,7 +48,8 @@ for filetype in filetypes:
 
     # Make backup!
     print 'Copying ', files['oldmsid'].abs, files['oldmsid'].abs + '.bak'
-    shutil.copy2(files['oldmsid'].abs, files['oldmsid'].abs + '.bak')
+    if not opt.dry_run:
+        shutil.copy2(files['oldmsid'].abs, files['oldmsid'].abs + '.bak')
 
     # Read the list of archive files that were used to build the MSID h5 files
     db = Ska.DBI.DBI(dbi='sqlite', server=files['archfiles'].abs)
@@ -47,17 +62,26 @@ for filetype in filetypes:
     file1s = archfiles[1:][overlaps]
 
     # Open the TIME.h5 file for this content type 
-    h5 = tables.openFile(files['oldmsid'].abs, mode='a')
+    h5 = tables.openFile(files['oldmsid'].abs, mode=('r' if opt.dry_run else 'a'))
 
     # Iterate through overlapping files and set bad quality in TIME for rows
     # coming from the file with lower revision
     for file0, file1 in zip(file0s, file1s):
-        t0 = file0.tstart
-        badfile = file0 if file0.revision < file1.revision else file1
-        print file0.filename, file0.revision, file1.revision, \
-              '%9d %9d %9d %9d' % (file0.tstart, file0.tstop, file1.tstart, file1.tstop), \
-              'Setting TIME rows', badfile.rowstart, badfile.rowstop, 'to bad quality'
-        h5.root.quality[badfile.rowstart:badfile.rowstop] = True
+        if opt.overwrite_partial:
+            # Effectively overwrite the file0 values with file1 values
+            # for the file0 rows from file1['tstart'] to file0['tstop']
+            times = h5.root.data[file0['rowstart']:file0['rowstop']]
+            bad_rowstart = numpy.searchsorted(times, file1['tstart']) + file0['rowstart']
+            bad_rowstop = file0['rowstop']
+        else:
+            badfile = file0 if file0['revision'] < file1['revision'] else file1
+            bad_rowstart = badfile['rowstart']
+            bad_rowstop = badfile['rowstop']
+        print file0['filename'], file0['revision'], file1['revision'], \
+              '%9d %9d %9d %9d' % (file0['tstart'], file0['tstop'], file1['tstart'], file1['tstop']), \
+              'Setting TIME rows', bad_rowstart, bad_rowstop, 'to bad quality'
+        if not opt.dry_run:
+            h5.root.quality[bad_rowstart:bad_rowstop] = True
 
     h5.close()
